pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tune"
version = "0.0.1"
description = "AI-powered bioinformatics data analysis platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "procrastinate[psycopg]>=2.10.0",
    "psycopg[async,binary]>=3.1.0",
    "pgvector>=0.3.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "watchdog>=4.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "PyYAML>=6.0.0",
    "anthropic>=0.30.0",
    "openai>=1.35.0",
    "python-multipart>=0.0.9",
    "psutil>=6.0.0",
    "aiofiles>=23.2.0",
]

[project.scripts]
tune = "tune.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["tune"]

[tool.hatch.build.targets.wheel.shared-data]
"tune/frontend" = "tune/frontend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
markers = [
    "slow: integration or real-data validation tests that are not part of the default fast suite",
]
